+++ Packages/com.anomalybound.jamforge/Runtime/Services/Log4Net/Abstractions/ILog.cs
@@ -1,4 +1,0 @@
-namespace JamForge.Log4Net
-{
-    public interface ILog : log4net.ILog { }
-}