+++ Packages/com.anomalybound.jamforge/Runtime/Services/Logging/Appenders/Formatter/LogFormatter.cs
@@ -1,0 +1,9 @@
+namespace JamForge.Logging
+{
+    public delegate string LogFormatter(JamLogger logger, JamLogLevel logLevel, string message);
+
+    public interface ILogFormatter
+    {
+        string FormatMessage(JamLogger logger, JamLogLevel logLevel, string message);
+    }
+}