+++ Packages/com.anomalybound.jamforge/Runtime/Services/StateMachine/Implements/StateMachineRunner.cs
@@ -79,20 +79,20 @@ namespace JamForge.StateMachine
         {
             if (States.TryAdd(index, state)) { return; }
 
-            Jam.Logger.E($"State machine already contains state with index {index}");
+            InternalLog.E($"State machine already contains state with index {index}");
         }
 
         public void AddTransition(TIndex from, TIndex to, Func<TState, bool> condition)
         {
             if (!States.TryGetValue(from, out _))
             {
-                Jam.Logger.E($"State machine does not contain state with index {from}");
+                InternalLog.E($"State machine does not contain state with index {from}");
                 return;
             }
 
             if (!States.TryGetValue(to, out _))
             {
-                Jam.Logger.E($"State machine does not contain state with index {to}");
+                InternalLog.E($"State machine does not contain state with index {to}");
                 return;
             }